+++ src/Getter.h
@@ -24,8 +24,8 @@ class Getter : public TObject {
 
   Getter() = default;
 
-  float GetCentrality(float value) const;
-  float GetCentrality(float xvalue, float yvalue) const;
+  double GetCentrality(double value) const;
+  double GetCentrality(double xvalue, double yvalue) const;
 
   void SetBorders(const std::vector<double> &borders) {
     borders_ = TAxis(borders.size() - 1, &(borders[0]));
@@ -33,19 +33,19 @@ class Getter : public TObject {
   }
 
   const TAxis &GetBorders() const { return borders_; };
-  const std::vector<float> &GetRanges() const { return ranges_; };
+  const std::vector<double> &GetRanges() const { return ranges_; };
 
-  void SetRanges(const std::vector<float> &ranges) { ranges_ = ranges; }
+  void SetRanges(const std::vector<double> &ranges) { ranges_ = ranges; }
   void IsSpectator(bool is = true) { isspectator_ = is; }
 
-  void AddBorder2D(const std::array<float, 2> &border2D) {
+  void AddBorder2D(const std::array<double, 2> &border2D) {
     borders2d_.push_back(border2D);
     if (!isinitialized2D_) isinitialized2D_ = true;
   }
 
-  const std::vector<std::array<float, 2>> &GetBorders2D() const { return borders2d_; };
+  const std::vector<std::array<double, 2>> &GetBorders2D() const { return borders2d_; };
 
-  void SetMax(float x, float y) {
+  void SetMax(double x, double y) {
     xmax_ = x;
     ymax_ = y;
   }
@@ -55,11 +55,11 @@ class Getter : public TObject {
  private:
 
   TAxis borders_;
-  std::vector<std::array<float, 2>> borders2d_{};
-  std::vector<float> ranges_{};
+  std::vector<std::array<double, 2>> borders2d_{};
+  std::vector<double> ranges_{};
 
-  float xmax_{1.};
-  float ymax_{1.};
+  double xmax_{1.};
+  double ymax_{1.};
 
   bool isspectator_{false};
   bool isinitialized_{false};
